requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=1.10.8
starlette>=0.27.0

sqlalchemy>=1.4.46
psycopg2-binary>=2.9.6

redis>=4.5.5

hashlib>=20081119

python-dotenv>=1.0.0

validators>=0.20.0

hvac>=1.1.0

requests>=2.30.0
